Split into UI components and services
