+++ scripts/services/search/constants.js
@@ -1,0 +1,6 @@
+export const RESULTS_LIMIT = 50;
+export const BASE_URL = document.querySelector("meta[name='baseurl']").content;
+
+export const SEARCHGOV_ENDPOINT = document.querySelector("meta[name='searchgov_endpoint']").content;
+export const SEARCHGOV_AFFILIATE = document.querySelector("meta[name='searchgov_affiliate']").content;
+export const SEARCHGOV_ACCESS_KEY = document.querySelector("meta[name='searchgov_access_key']").content;
